+++ test/check_test.dart
@@ -16,7 +16,7 @@ library quiver.check_test;
 
 import 'package:quiver/check.dart';
 import 'package:matcher/matcher.dart';
-import 'package:unittest/unittest.dart';
+import 'package:test/test.dart';
 
 main() {
   group('checkArgument', () {