+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#### Unreleased
+   * Switch from unittest to test.
+
 #### 0.21.3 - 2015-03-03
    * Bugfix: fixed return type on some methods (e.g. `where` of `Iterable`s
      returned by Multimap.